Refactor integration test for better reusability
