+++ components/event-bus/test/integration/publish_push_test.go
@@ -1,44 +1,20 @@
-package publish_push_test
+package integration
 
 import (
-	"encoding/json"
-	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"net/http/httptest"
 	"os"
-	"strings"
 	"testing"
-	"time"
-
-	api "github.com/kyma-project/kyma/components/event-bus/api/publish"
 	publishapp "github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish/application"
 	pushapp "github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-push/application"
-	"github.com/kyma-project/kyma/components/event-bus/generated/push/clientset/versioned/fake"
-	"github.com/kyma-project/kyma/components/event-bus/generated/push/informers/externalversions/eventing.kyma.cx/v1alpha1"
 	"github.com/kyma-project/kyma/components/event-bus/internal/common"
 	"github.com/kyma-project/kyma/components/event-bus/internal/publish"
 	"github.com/kyma-project/kyma/components/event-bus/internal/push/actors"
 	"github.com/kyma-project/kyma/components/event-bus/internal/push/opts"
 	"github.com/kyma-project/kyma/components/event-bus/test/util"
-	"github.com/nats-io/nats-streaming-server/server"
-	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/tools/cache"
-)
-
-const (
-	clusterID               = "kyma-nats-streaming"
-	eventType               = "test-publish-push-success"
-	eventTypeVersion        = "v1"
-	sourceIDV1              = "test.local.kyma.commerce.ec"
-	eventDataV1             = "test-event-1"
-	sourceIDV2              = "test.local.kyma.commerce.ec"
-	eventDataV2             = "test-event-2"
-	publishServerStatusPath = "/v1/status/ready"
-	headerKymaTopic         = "kyma-topic"
 )
 
 var (
@@ -49,14 +25,6 @@ var (
 	subscriptionsSupervisor1 *actors.SubscriptionsSupervisor
 	subscriptionsSupervisor2 *actors.SubscriptionsSupervisor
 )
-
-func startNats() (*server.StanServer, error) {
-	return server.RunServer(clusterID)
-}
-
-func stopNats(stanServer *server.StanServer) {
-	stanServer.Shutdown()
-}
 
 func TestMain(m *testing.M) {
 
@@ -76,7 +44,7 @@ func TestMain(m *testing.M) {
 
 	pushOpts := opts.DefaultOptions
 	pushOpts.NatsStreamingClusterID = clusterID
-	pushApplication := pushapp.NewPushApplication(&pushOpts, newFakeInformer())
+	pushApplication := pushapp.NewPushApplication(&pushOpts, newFakeInformer2())
 	subscriptionsSupervisor1 = pushApplication.SubscriptionsSupervisor
 	subscription1 := util.NewSubscription("test-sub", metav1.NamespaceDefault, subscriberServerV1.URL+util.SubServer1EventsPath, eventType, eventTypeVersion,
 		sourceIDV1)
@@ -127,102 +95,18 @@ func Test_Subscriber_Status(t *testing.T) {
 	verifyStatusCode(res2, http.StatusOK, t)
 }
 
-func makePayload(sourceID, eventType, eventTypeVersion, eventData string) string {
-	return fmt.Sprintf(`{"source-id": "%s", "event-type": "%s","event-type-version": "%s","event-time": "2018-11-02T22:08:41+00:00","data": "%s"}`,
-		sourceID, eventType, eventTypeVersion, eventData)
-}
-
 func Test_Publish_Push_Request(t *testing.T) {
 	{
 		payloadV1 := makePayload(sourceIDV1, eventType, eventTypeVersion, eventDataV1)
-		res, err := http.Post(publishServer.URL+"/v1/events", "application/json", strings.NewReader(payloadV1))
-		checkIfError(err, t)
-		verifyStatusCode(res, 200, t)
-		log.Print(res)
-		respObj := &api.PublishResponse{}
-		body, err := ioutil.ReadAll(res.Body)
-		defer res.Body.Close()
-		err = json.Unmarshal(body, &respObj)
-		assert.NotNil(t, respObj.EventID)
-		assert.NotEmpty(t, respObj.EventID)
-		log.Printf("%v", respObj)
-
-		var ok bool
-		for i := 0; i < 10; i++ {
-			time.Sleep(1 * time.Second)
-			res, err := http.Get(subscriberServerV1.URL + util.SubServer1ResultsPath)
-			assert.Nil(t, err)
-			body, err := ioutil.ReadAll(res.Body)
-			var resp string
-			json.Unmarshal(body, &resp)
-			res.Body.Close()
-			if len(resp) == 0 {
-				continue
-			}
-			assert.Equal(t, eventDataV1, resp)
-			ok = true
-			break
-		}
-		assert.True(t, ok)
+		publishEvent(t, publishServer.URL, payloadV1)
+		verifyEndpointReceivedEvent(t, subscriberServerV1.URL+util.SubServer1ResultsPath, eventDataV1)
 	}
 
 	{
 		payloadV2 := makePayload(sourceIDV2, eventType, eventTypeVersion, eventDataV2)
-		res, err := http.Post(publishServer.URL+"/v1/events", "application/json", strings.NewReader(payloadV2))
-		checkIfError(err, t)
-		verifyStatusCode(res, 200, t)
-		log.Print(res)
-		respObj := &api.PublishResponse{}
-		body, err := ioutil.ReadAll(res.Body)
-		defer res.Body.Close()
-		err = json.Unmarshal(body, &respObj)
-		assert.NotNil(t, respObj.EventID)
-		assert.NotEmpty(t, respObj.EventID)
-		log.Printf("%v", respObj)
+		publishEvent(t, publishServer.URL, payloadV2)
+		verifyEndpointReceivedEvent(t, subscriberServerV2.URL+util.SubServer2ResultsPath, eventDataV2)
 
-		var ok bool
-		for i := 0; i < 10; i++ {
-			time.Sleep(1 * time.Second)
-			res, err := http.Get(subscriberServerV2.URL + util.SubServer2ResultsPath)
-			assert.Nil(t, err)
-			body, err := ioutil.ReadAll(res.Body)
-			var resp string
-			json.Unmarshal(body, &resp)
-			res.Body.Close()
-			if len(resp) == 0 {
-				continue
-			}
-			assert.Equal(t, eventDataV2, resp)
-			ok = true
-			break
-		}
-		assert.True(t, ok)
-	}
-}
-
-func newFakeInformer() cache.SharedIndexInformer {
-	sub := util.NewSubscription(
-		"test-sub",
-		metav1.NamespaceDefault,
-		subscriberServerV1.URL+util.SubServer1EventsPath,
-		eventType,
-		eventTypeVersion,
-		sourceIDV1)
-	clientSet := fake.NewSimpleClientset(sub)
-	informer := v1alpha1.NewSubscriptionInformer(clientSet, metav1.NamespaceAll, 0, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-	informer.GetIndexer().Add(sub)
-	return informer
-}
-
-func verifyStatusCode(res *http.Response, expectedStatusCode int, t *testing.T) {
-	if res.StatusCode != expectedStatusCode {
-		t.Errorf("Status code is wrong, have: %d, want: %d", res.StatusCode, expectedStatusCode)
-	}
-}
-
-func checkIfError(err error, t *testing.T) {
-	if err != nil {
-		t.Fatal(err)
 	}
 }
 
@@ -237,38 +121,11 @@ func Test_pushRequestShouldNotIncludeKymaTopicHeader(t *testing.T) {
 	subscription1 := util.NewSubscription("test-sub-1", metav1.NamespaceDefault, subscriberServerV1.URL+util.SubServer1EventsPath, eventType, eventTypeVersion,
 		sourceIDV1)
 	subscriptionsSupervisor1.StartSubscriptionReq(subscription1, requestProvider)
-	{
-		payloadV1 := makePayload(sourceIDV1, eventType, eventTypeVersion, eventDataV1)
-		res, err := http.Post(publishServer.URL+"/v1/events", "application/json", strings.NewReader(payloadV1))
-		checkIfError(err, t)
-		verifyStatusCode(res, 200, t)
-		log.Print(res)
-		respObj := &api.PublishResponse{}
-		body, err := ioutil.ReadAll(res.Body)
-		defer res.Body.Close()
-		err = json.Unmarshal(body, &respObj)
-		assert.NotNil(t, respObj.EventID)
-		assert.NotEmpty(t, respObj.EventID)
-		log.Printf("%v", respObj)
 
-		var ok bool
-		for i := 0; i < 20; i++ {
-			time.Sleep(1 * time.Second)
-			res, err := http.Get(subscriberServerV1.URL + util.SubServer1ResultsPath)
-			assert.Nil(t, err)
-			body, err := ioutil.ReadAll(res.Body)
-			var resp string
-			json.Unmarshal(body, &resp)
-			res.Body.Close()
-			if len(resp) == 0 {
-				continue
-			}
-			assert.Equal(t, eventDataV1, resp)
-			ok = true
-			break
-		}
-		assert.True(t, ok)
-	}
+	payloadV1 := makePayload(sourceIDV1, eventType, eventTypeVersion, eventDataV1)
+	publishEvent(t, publishServer.URL, payloadV1)
+
+	verifyEndpointReceivedEvent(t, subscriberServerV1.URL+util.SubServer1ResultsPath, eventDataV1)
 
 	if pushRequest == nil {
 		t.Fatal("push request should not be nil")
@@ -291,45 +148,9 @@ func Test_sameSubjectSubscribersInDifferentNamespacesShouldReceiveEventsOfThatSu
 
 	// publish one event
 	payload := makePayload(sourceIDV1, eventType, eventTypeVersion, eventDataV1)
-	publishEvent(t, payload)
+	publishEvent(t, publishServer.URL, payload)
 
 	// verify that both subscribers received the event
 	verifyEndpointReceivedEvent(t, subscriberServerV1.URL+util.SubServer1ResultsPath, eventDataV1)
 	verifyEndpointReceivedEvent(t, subscriberServerV2.URL+util.SubServer2ResultsPath, eventDataV1)
-}
-
-func publishEvent(t *testing.T, payload string) {
-	res, err := http.Post(publishServer.URL+"/v1/events", "application/json", strings.NewReader(payload))
-	checkIfError(err, t)
-	verifyStatusCode(res, 200, t)
-	log.Print(res)
-
-	respObj := &api.PublishResponse{}
-	body, err := ioutil.ReadAll(res.Body)
-	defer res.Body.Close()
-
-	err = json.Unmarshal(body, &respObj)
-	assert.NotNil(t, respObj.EventID)
-	assert.NotEmpty(t, respObj.EventID)
-	log.Printf("%v", respObj)
-}
-
-func verifyEndpointReceivedEvent(t *testing.T, endpoint, data string) {
-	var ok bool
-	for i := 0; i < 20; i++ {
-		time.Sleep(1 * time.Second)
-		res, err := http.Get(endpoint)
-		assert.Nil(t, err)
-		body, err := ioutil.ReadAll(res.Body)
-		var resp string
-		json.Unmarshal(body, &resp)
-		res.Body.Close()
-		if len(resp) == 0 {
-			continue
-		}
-		assert.Equal(t, data, resp)
-		ok = true
-		break
-	}
-	assert.True(t, ok)
 }
